% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotations.R
\name{add_annotations}
\alias{add_annotations}
\title{A function to add annotations to a table of gene counts.}
\usage{
add_annotations(object, reference = geneID.details, variables = NULL)
}
\arguments{
\item{object}{A table of gene counts (rows: genes, columns:samples).}

\item{reference}{A reference table with the annotations. Default: geneID.details.}

\item{variables}{Annotations (columns) from the reference to add, i.e. (gene symbol, gene_length, description).}
}
\description{
A function to add annotations to a table of gene counts.
}
