% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{This function annotates a column of transcripts or gene IDs (ENSEMBL) with information of the Biomart.
If transcript IDs are provided, they are also annotated with information of the genes to which they belong.}
\usage{
get_annotations(ensembl_ids, mode = "genes", version = "", format = "csv")
}
\arguments{
\item{ensembl_ids}{The column of transcripts to be used as input.}

\item{mode}{To specify the IDs provided, between "transcripts" or "genes". Default = genes.}

\item{version}{This function can use the version 103 or the current version of the Biomart. Default = Current.}

\item{format}{The output will be a table named tx2gene in .csv or .xlsx formats. Default = csv.}
}
\description{
The Gene information added include:
\itemize{
\item Gene ENSEMBL ID, HGNC Symbol, Description, Biotype and Chromosome.
\item Gene start, end and length
}
}
