% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_PCA.R
\name{nice_PCA}
\alias{nice_PCA}
\title{Function to make nice PCA plots}
\usage{
nice_PCA(
  object,
  PCs = c(1, 2),
  ntop = 200,
  returnData = FALSE,
  variables = c(fill = "VarFill", shape = "VarShape"),
  legend_names = c(fill = "Sample Type", shape = "Library"),
  size = 7,
  alpha = 1,
  colors = NULL,
  shapes = NULL,
  legend_title = 20,
  legend_elements = 16,
  legend_pos = c(0.8, 0.8, "right"),
  labels = NULL,
  name_tags = NULL
)
}
\arguments{
\item{object}{A DEseq object already transformed with the variance stabilizing or rlog transformations.}

\item{PCs}{A vector indicating the two Principal Components to plot. Default: c(1,2).}

\item{ntop}{Number of top genes to use for principal components, selected by highest row variance.}

\item{returnData}{Indicates if the function should return the data (TRUE) or the plot (FALSE). Default: FALSE.}

\item{variables}{To indicate the variables to be used as Shape and Fill of the markers.}

\item{legend_names}{The names to be used for the legend of the Shape and Fill.}

\item{size}{Size of the marker. Default: 7.}

\item{alpha}{Transparency of the marker, which goes from 0 (transparent) to 1 (no transparent). Default: 1.}

\item{colors}{Vector of colors to be used for the categories of the variable assigned as Marker Fill.}

\item{shapes}{Vector of shapes to be used for the categories of the variable assigned as Marker Shape.}

\item{legend_title}{Font of the legend title. Default: 20.}

\item{legend_elements}{Font of the elements of the legend Default: 16.}

\item{legend_pos}{Position of the legend in the plot. Default: c(0.80, 0.80, "right").}

\item{labels}{A vector containing the variable to be used as labels (name inside the marker), and the label size. Example: c(var = "patient", size = 2). Default: NULL (no labels).}

\item{name_tags}{A vector containing the variable to be used as name tags (name outside the marker), tag size, minimum distance in order to add an arrow connecting the tag and the marker, and minimum distance from the tag and the center of the marker. Example: c(var = "label", size = 2, minlen = 2, box = 0.6). Default: NULL (no name tags).}
}
\description{
This was inspired on the plotPCA function from DESeq2, made by Wolfgang Huber
But including some improvements made by David Requena. Now it allows:
\itemize{
\item To choose which PCs to plot.
\item To use one or two features to represent as the fill or shape of the markers.
\item To provide the colors, shapes and fonts.
}
}
