% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{tpm}
\alias{tpm}
\title{Function to calculate the TPMs from a table of raw gene counts.}
\usage{
tpm(raw_counts, gene_lengths)
}
\arguments{
\item{raw_counts}{A table with the gene counts.}

\item{gene_lengths}{A column with the gene lengths.}
}
\description{
TPM: Transcript per million. See https://www.biostars.org/p/273537/
The input table is numeric:
\itemize{
\item The row names are the gene identifiers (ensembl ID).
\item The column names represent the samples.
The gene lengths are in a column of a dataframe with the same row order.
}
}
