% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{This function annotates a column of transcripts IDs (ENSEMBL) with information of the genes to which they belong.}
\usage{
get_annotations(tx_ensembl_ids, version = "103", format = "csv")
}
\arguments{
\item{tx_ensembl_ids}{The column of transcripts to be used as input.}

\item{version}{This function can use the version 103 or the current version of the Biomart. Default = 103.}

\item{format}{The output will be a table named tx2gene in .csv or .xlsx formats. Default = csv.}
}
\description{
The Gene information added include:
\itemize{
\item ENSEMBL ID
\item HGNC Symbol and description
\item Start and End position
\item Gene length
}
}
